+++ EventService/src/main/java/com/enfec/sb/eventapi/EventApiApplication.java
@@ -1,4 +1,4 @@
-package com.enfec.sb.organizerapi;
+package com.enfec.sb.eventapi;
 
 
 import org.springframework.boot.SpringApplication;
