+++ EventService/src/main/java/com/enfec/sb/eventapi/model/EventTable.java
@@ -1,4 +1,4 @@
-package com.enfec.sb.organizerapi.model;
+package com.enfec.sb.eventapi.model;
 
 import java.sql.Timestamp;
 import java.util.Date;
