+++ EventService/src/main/resources/assembly/assembly.xml
